Code Quality: Move some train stuff into separate files, some comments.
[ci skip]
