+++ openBVE/OpenBve/OpenBve.csproj
@@ -129,6 +129,9 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
+    <Compile Include="Simulation\TrainManager\Doors.cs" />
+    <Compile Include="Simulation\TrainManager\Passengers.cs" />
+    <Compile Include="Simulation\TrainManager\Station.cs" />
     <Compile Include="System\Functions\CrashHandler.cs" />
     <Compile Include="System\Functions\FolderBrowserDialog.cs" />
     <Compile Include="System\Input\Commands.cs" />
@@ -317,8 +320,8 @@ if %25textTemplatingPath%25=="\Microsoft Shared\TextTemplating\$(VisualStudioVer
     <PreBuildEvent Condition=" '$(OS)' == 'Unix' ">$(ProjectDir)Properties/AssemblyInfo.sh</PreBuildEvent>
   </PropertyGroup>
   <PropertyGroup>
-  <!-- Copy in dependancies and child projects -->
-	<PostBuildEvent Condition=" '$(OS)' != 'Unix' ">
+    <!-- Copy in dependancies and child projects -->
+    <PostBuildEvent Condition=" '$(OS)' != 'Unix' ">
 		xcopy "$(ProjectDir)Data\*.*" "$(TargetDir)Data" /Y /I /E /C 
 		xcopy "$(ProjectDir)..\..\Dependencies\AtsPluginProxy.dll" "$(TargetDir)AtsPluginProxy.dll" /Y /I /C
 		xcopy "$(ProjectDir)..\OpenBveAts\bin\$(Configuration)\*.*" "$(TargetDir)" /Y /I /E /C
@@ -328,7 +331,7 @@ if %25textTemplatingPath%25=="\Microsoft Shared\TextTemplating\$(VisualStudioVer
 		xcopy "$(ProjectDir)..\Texture.BmpGifJpegPngTiff\bin\$(Configuration)\*.*" "$(TargetDir)" /Y /I /E /C
 	</PostBuildEvent>
     <!-- NOTE: Whilst AtsPluginProxy.dll is Windows specific, we still want to copy it in the auto-generated nightly builds -->
-	<PostBuildEvent Condition=" '$(OS)' == 'Unix' ">
+    <PostBuildEvent Condition=" '$(OS)' == 'Unix' ">
 		cp -rf "$(ProjectDir)Data" "$(ProjectDir)../../Dependencies/OpenTK.dll.config" "$(ProjectDir)../../Dependencies/AtsPluginProxy.dll" "$(ProjectDir)../OpenBveAts/bin/$(Configuration)/Data" "$(ProjectDir)../Sound.Flac/bin/$(Configuration)/Data" "$(ProjectDir)../Sound.RiffWave/bin/$(Configuration)/Data" "$(ProjectDir)../Texture.Ace/bin/$(Configuration)/Data" "$(ProjectDir)../Texture.BmpGifJpegPngTiff/bin/$(Configuration)/Data" "$(TargetDir)"
 	</PostBuildEvent>
   </PropertyGroup>
